doesn't allow deletion for non existing transactions

+++ Commands/TransactionCommands/DeleteTransactionCommand.cs
@@ -14,10 +14,26 @@ public class DeleteTransactionCommand : Command
     {
         Console.Clear();
 
+        PostgresTransactionManager postgresTransactionManager = new(connection);
+
+        List<Transaction> transactions = await postgresTransactionManager.GetAllTransactions();
+
+        foreach (Transaction transaction in transactions)
+        {
+            Console.WriteLine(transaction);
+        }
+
         Console.Write("What transaction do you wish to delete? ");
         int transactionToDelete = int.Parse(Console.ReadLine()!);
 
-        PostgresTransactionManager postgresTransactionManager = new(connection);
+        int rowCount = await postgresTransactionManager.GetTransactionsCount();
+
+        if (transactionToDelete >= rowCount)
+        {
+            Console.WriteLine("Transaction does not exist.");
+            PressKeyToContinue.Execute();
+            return;
+        }
 
         await postgresTransactionManager.DeleteTransaction(transactionToDelete);
